+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 ### Changes
 * Increased Sprint Pact movement speed (x1.15 -> x1.2)
+* Increased Adrenaline bonus mining speed and attack speed (exponential -> linear) but lowered bonus mining speed (2.5x -> 2x Efficiency enchantment bonus mining speed)
 
 ### Fixes
 * Sprint Pact and Curse of Walking no longer prevent players from swimming